+++ driver/lib/desired-caps.ts
@@ -5,9 +5,6 @@ export const desiredCapConstraints = {
   avd: {
     isString: true,
   },
-  maxRetryCount: {
-    isNumber: true,
-  },
   platformName: {
     inclusionCaseInsensitive: [
       'iOS',
@@ -15,9 +12,6 @@ export const desiredCapConstraints = {
     ],
     isString: true,
     presence: true,
-  },
-  retryBackoffTime: {
-    isNumber: true,
   },
   udid: {
     isString: true,